+++ src/main/kotlin/com/yapp/lettie/api/timecapsule/service/reader/TimeCapsuleUserReader.kt
@@ -1,6 +1,6 @@
 package com.yapp.lettie.api.timecapsule.service.reader
 
-import com.yapp.lettie.api.timecapsule.service.dto.RecipientRow
+import com.yapp.lettie.domain.timecapsule.dto.RecipientRow
 import com.yapp.lettie.common.error.ErrorMessages
 import com.yapp.lettie.common.exception.ApiErrorException
 import com.yapp.lettie.domain.timecapsule.entity.TimeCapsuleUser
