+++ src/main/kotlin/com/yapp/lettie/domain/timecapsule/dto/RecipientRow.kt
@@ -1,4 +1,4 @@
-package com.yapp.lettie.api.timecapsule.service.dto
+package com.yapp.lettie.domain.timecapsule.dto
 
 data class RecipientRow(
     val capsuleId: Long,
