+++ src/main/kotlin/com/yapp/lettie/domain/timecapsule/repository/TimeCapsuleUserCustomRepository.kt
@@ -1,6 +1,6 @@
 package com.yapp.lettie.domain.timecapsule.repository
 
-import com.yapp.lettie.api.timecapsule.service.dto.RecipientRow
+import com.yapp.lettie.domain.timecapsule.dto.RecipientRow
 
 interface TimeCapsuleUserCustomRepository {
     fun findRecipientsByCapsuleIds(capsuleIds: List<Long>): List<RecipientRow>
