+++ src/main/kotlin/com/yapp/lettie/domain/timecapsule/repository/TimeCapsuleUserCustomRepositoryImpl.kt
@@ -2,7 +2,7 @@ package com.yapp.lettie.domain.timecapsule.repository
 
 import com.querydsl.core.types.Projections
 import com.querydsl.jpa.impl.JPAQueryFactory
-import com.yapp.lettie.api.timecapsule.service.dto.RecipientRow
+import com.yapp.lettie.domain.timecapsule.dto.RecipientRow
 import com.yapp.lettie.domain.timecapsule.entity.QTimeCapsuleUser
 import com.yapp.lettie.domain.user.entity.QUser
